Apply RequestDTO
